Migrated Sign In / Sign Up logic to single screen
