+++ lib/auth.dart
